pyproject.toml
[tool.poetry]
name = "fatigue-ml"
version = "0.1.0"
description = "Top-tier machine learning pipeline for cognitive fatigue detection using time-series biosignal data."
authors = ["Anikait Lakhotia <alokhoti@uwaterloo.ca>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"

# Core numerical and ML stack
numpy = "^1.26"
scipy = "^1.12"
pandas = "^2.1"
scikit-learn = "^1.4"

# Visualization
matplotlib = "^3.8"
seaborn = "^0.13"

# Deep learning
torch = "^2.2"
torchaudio = "^2.2"
pytorch-lightning = "^2.2"

# Configuration, experiment management, and tracking
hydra-core = "^1.3"
mlflow = "^2.12"
wandb = "^0.17"

# Signal processing & dimensionality reduction
mne = "^1.6"
umap-learn = "^0.5"
tqdm = "^4.66"

# Data quality & validation
great-expectations = "^0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-cov = "^5.0"
black = "^24.3"
isort = "^5.13"
ruff = "^0.3"
mypy = "^1.9"
pre-commit = "^3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# -----------------------------
# Tooling Configuration
# -----------------------------

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | \.pytest_cache
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
skip_gitignore = true

[tool.ruff]
line-length = 88
select = ["E", "F", "B", "I", "UP"]
ignore = ["E203", "W503"]
target-version = "py311"
exclude = ["build", "dist", ".venv"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra -q --disable-warnings"

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
